+++ src/ts/uimanager.ts
@@ -9,6 +9,7 @@ import {ArrayUtils} from './arrayutils';
 import {BrowserUtils} from './browserutils';
 import { TimelineMarker, UIConfig } from './uiconfig';
 import { PlayerAPI, PlayerEventCallback, PlayerEventBase, PlayerEvent, AdEvent } from 'bitmovin-player';
+import { VolumeController } from './volumecontroller';
 
 export interface InternalUIConfig extends UIConfig {
   events: {
@@ -17,6 +18,7 @@ export interface InternalUIConfig extends UIConfig {
      */
     onUpdated: EventDispatcher<UIManager, void>;
   };
+  volumeController: VolumeController;
 }
 
 /**
@@ -124,6 +126,7 @@ export class UIManager {
     }
 
     this.player = player;
+    this.managerPlayerWrapper = new PlayerWrapper(player);
 
     // ensure that at least the metadata object does exist in the uiconfig
     uiconfig.metadata = uiconfig.metadata ? uiconfig.metadata : {};
@@ -135,8 +138,8 @@ export class UIManager {
       events: {
         onUpdated: new EventDispatcher<UIManager, void>(),
       },
+      volumeController: new VolumeController(this.managerPlayerWrapper.getPlayer()),
     };
-    this.managerPlayerWrapper = new PlayerWrapper(player);
 
     /**
      * Gathers configuration data from the UI config and player source config and creates a merged UI config