+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - CSS selector to default elements
 - `VolumeController` to control and manage volume and mute state by multiple `Component`s in a single place
+- `disable()` / `enable()` functionality to `Component`s
+- Preventing click event on a disabled `Button`
 - Ads UI support for player v8
 - `UIConditionContext#adRequiresUi` to distinguish if the current ad requires an external UI